refactor(server): split runtime worker child tests
